refactor(cli): extract commands into cli/commands/ per ADR-011
Implements plan 040 Wave 3.

- Splits 563-LOC cli/index.ts into 8 focused command modules
  (claim, db, entity, export, import, link, note, search) plus
  a shared context module
- Each module exports a registerXxxCommand(program, ctx) function
  accepting an injected CommandContext — no global state
- cli/index.ts shrinks to 129 LOC (entry point only, mounts all
  commands)
- New `pnpm run cli -- import <file>` command supports both JSON
  (validated against KnowledgeStudioExport v1.0) and Markdown
  (gray-matter frontmatter)
- New `pnpm run cli -- export -f pdf` uses the @react-pdf/renderer
  pipeline added in the previous commit
- 3 new CLI tests verify all command groups register correctly
  (395 tests total)

Per AGENTS.md hard rule: all files under 200 LOC (largest is 151).

Closes: #289 (cli/commands part)
Refs: ADR-011
